requirements.txt
absl-py==2.1.0
blinker==1.8.1
cachetools==5.3.3
certifi==2019.11.28
chardet==3.0.4
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
dbus-python==1.2.16
distro==1.4.0
filelock==3.13.4
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.3.1
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.62.1
idna==2.8
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==24.0
pandas==2.0.3
pillow==10.3.0
protobuf==5.26.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyGObject==3.36.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.22.0
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.10.1
six==1.14.0
ssh-import-id==5.10
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.25.8
Werkzeug==3.0.2
zipp==3.18.1
